pyproject.toml
[project]
name = "shipping-erp-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "유선혁",email = "tjsgur2399@icloud.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "multipart (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "shipping_erp_fastapi", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
